Fix entity module CI regression fixtures
Set generated-model fixture state explicitly so schema defaults do not
change preparation ordering or attribution preservation assertions. Cover
invalid-tag rejection for every lifecycle status and an explicit null.

Use a long lookup counter to remove CodeQL's narrowing-conversion finding.
Record the CI artifact reproduction, full service test results, and the
downstream Collate Java migration still required by its build checks.

+++ openmetadata-service/src/test/java/org/openmetadata/service/entity/bulk/EntityBulkPreparationTest.java
@@ -141,7 +141,7 @@ class EntityBulkPreparationTest {
     final List<UUID> loaded = new ArrayList<>();
     final List<String> visited = new ArrayList<>();
     final EntityBulkPreparation<Chart> preparation;
-    int reads;
+    long reads;
     int clears;
     int maximumActive;
     boolean failRead;
